fix(router): expose inject function for page endpoint URL
