+++ packages/router/src/index.ts
@@ -13,3 +13,4 @@ export { PageServerLoad, LoadResult } from './lib/route-types';
 export { injectLoad } from './lib/inject-load';
 export { getLoadResolver } from './lib/get-load-resolver';
 export { requestContextInterceptor } from './lib/request-context';
+export { injectRouteEndpointURL } from './lib/inject-route-endpoint-url';
