+++ packages/router/src/lib/route-config.ts
@@ -1,12 +1,12 @@
 import { inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import type { Route } from '@angular/router';
-import { injectBaseURL } from '@analogjs/router/tokens';
 import { firstValueFrom } from 'rxjs';
 
 import { RedirectRouteMeta, RouteConfig, RouteMeta } from './models';
 import { ROUTE_META_TAGS_KEY } from './meta-tags';
 import { ANALOG_PAGE_ENDPOINTS, ANALOG_META_KEY } from './endpoints';
+import { injectRouteEndpointURL } from './inject-route-endpoint-url';
 
 export function toRouteConfig(routeMeta: RouteMeta | undefined): RouteConfig {
   if (routeMeta && isRedirectRouteMeta(routeMeta)) {
@@ -33,34 +33,13 @@ export function toRouteConfig(routeMeta: RouteMeta | undefined): RouteConfig {
   routeConfig.resolve = {
     ...routeConfig.resolve,
     load: async (route) => {
-      const baseUrl = injectBaseURL();
+      const http = inject(HttpClient);
       const routeConfig = route.routeConfig as Route & {
         [ANALOG_META_KEY]: { endpoint: string; endpointKey: string };
       };
 
       if (ANALOG_PAGE_ENDPOINTS[routeConfig[ANALOG_META_KEY].endpointKey]) {
-        const { queryParams, fragment: hash, params, parent } = route;
-        const segment =
-          parent?.url.map((segment) => segment.path).join('/') || '';
-        const url = new URL(
-          '',
-          import.meta.env['VITE_ANALOG_PUBLIC_BASE_URL'] ||
-            baseUrl ||
-            (typeof window !== 'undefined' && window.location.origin
-              ? window.location.origin
-              : '')
-        );
-
-        url.pathname = `${
-          url.pathname.endsWith('/') ? url.pathname : url.pathname + '/'
-        }api/_analog${routeConfig[ANALOG_META_KEY].endpoint}`;
-        url.search = `${new URLSearchParams(queryParams).toString()}`;
-        url.hash = hash ?? '';
-
-        Object.keys(params).forEach((param) => {
-          url.pathname = url.pathname.replace(`[${param}]`, params[param]);
-        });
-        url.pathname = url.pathname.replace('**', segment);
+        const url = injectRouteEndpointURL(route);
 
         if (
           !!import.meta.env['VITE_ANALOG_PUBLIC_BASE_URL'] &&
@@ -69,7 +48,6 @@ export function toRouteConfig(routeMeta: RouteMeta | undefined): RouteConfig {
           return (globalThis as any).$fetch(url.pathname);
         }
 
-        const http = inject(HttpClient);
         return firstValueFrom(http.get(`${url.href}`));
       }
